A rudimentary implementation of the Stack class using pointers.
Notes: The getMiddle() function currently has a complexity of O(n).
This can be alleviated by keeping two pointers in the Stack class instead
of one and changing the push method to push on alternate stacks on
consecutive calls. This way, one of the stacks always has the top element.
We would also need to track which stack last received a 'push'.